Using Parcelable
The Bundle object which is used to pass data to Android components is a key/value store for specialized objects.
It is similar to a Map but can only contain these specialized objects(Primitive Data Types).
Wou can place the following objects types into a Bundle:
If we need to pass ours customer objects via a Bundle, we should implement the Parcelable interface.

Signed-off-by: Solo <anqi.huang@outlook.com>